\name{gjrfit-package}
\alias{gjrfit-package}
\alias{gjrfit}
\docType{package}
\title{Generalized Joint Rankings Estimates}
\description{
Uses a transformation-retransformation approach to estimate parameters in a line
ar model with cluster correlated error.
}
\details{
\tabular{ll}{
Package: \tab gjrfit\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-04-03\cr
License: \tab GPL (version 2 or later)\cr
}
}
\author{
John Kloke

Maintainer John Kloke <kloke@biostat.wisc.edu>
}
\references{
Kloke (2015) In preparation
}
\keyword{ package }
